Add trait HasExecutionPlannedAt
